# src/main/resources/com/sonicle/webtop/calendar/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Calendar",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Calendar service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    },
    {
      "auth-apikey-bearer": [],
      "auth-apikey-username": []
    }
  ],
  "paths": {
    "/caldav/calendars": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "List all calendars",
        "operationId": "getDavCalendars",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DavCalendar"
                  }
                }
              }
            }
          }
        },
        "description": "List available Calendars."
      },
      "post": {
        "tags": [
          "dav"
        ],
        "summary": "Adds a new calendar",
        "operationId": "addDavCalendar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavCalendarNew"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCalendar"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Creates new Calendar."
      }
    },
    "/caldav/calendars/{calendarUid}": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Gets a single calendar",
        "operationId": "getDavCalendar",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCalendar"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Calendar not found",
            "content": {}
          }
        },
        "description": "Gets specified Calendar."
      },
      "put": {
        "tags": [
          "dav"
        ],
        "summary": "Updates a calendar",
        "operationId": "updateDavCalendar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavCalendarUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Calendar not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates specified Calendar."
      },
      "delete": {
        "tags": [
          "dav"
        ],
        "summary": "Deletes a calendar",
        "operationId": "deleteDavCalendar",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Calendar not found",
            "content": {}
          },
          "405": {
            "description": "Delete operation is not allowed",
            "content": {}
          }
        },
        "description": "Deletes specified Calendar."
      },
      "parameters": [
        {
          "name": "calendarUid",
          "in": "path",
          "description": "Calendar UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/caldav/calendars/{calendarUid}/objects": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "List all calendar objects belonging to a specific calendar",
        "operationId": "getDavCalObjects",
        "parameters": [
          {
            "name": "hrefs",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DavCalObject"
                  }
                }
              }
            }
          }
        },
        "description": "List all CalendarObjects of specified Calendar."
      },
      "post": {
        "tags": [
          "dav"
        ],
        "summary": "Adds a new calendar object",
        "operationId": "addDavCalObject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavCalObjectNew"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Creates new CalendarObject into specified Calendar."
      },
      "parameters": [
        {
          "name": "calendarUid",
          "in": "path",
          "description": "Calendar UID",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/caldav/calendars/{calendarUid}/objects/changes": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Get calendar object changes",
        "description": "Returns changed CalendarObjects (added/modified/deleted) since the specified syncToken. If token is not provided, the initial sync configuration will be returned.",
        "operationId": "getDavCalObjectsChanges",
        "parameters": [
          {
            "name": "syncToken",
            "in": "query",
            "description": "Marks changes starting point",
            "schema": {
              "type": "string"
            },
            "example": "20180416130954"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limits the number of returned results",
            "schema": {
              "type": "integer"
            },
            "example": 100
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCalObjectsChanges"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          }
        }
      },
      "parameters": [
        {
          "name": "calendarUid",
          "in": "path",
          "description": "Calendar UID",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/caldav/calendars/{calendarUid}/objects/{href}": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Get a single calendar object",
        "operationId": "getDavCalObject",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCalObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Object not found",
            "content": {}
          }
        },
        "description": "Gets specified CalendarObject."
      },
      "put": {
        "tags": [
          "dav"
        ],
        "summary": "Updates a calendar object",
        "operationId": "updateDavCalObject",
        "requestBody": {
          "content": {
            "text/calendar": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Object not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates specified CalendarObject."
      },
      "delete": {
        "tags": [
          "dav"
        ],
        "summary": "Deletes a calendar object",
        "operationId": "deleteDavCalObject",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Object not found",
            "content": {}
          }
        },
        "description": "Deletes specified CalendarObject."
      },
      "parameters": [
        {
          "name": "calendarUid",
          "in": "path",
          "description": "Calendar UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        },
        {
          "name": "href",
          "in": "path",
          "description": "CalObject reference URI",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "DavCalendar": {
        "type": "object",
        "description": "Bean for carry calendar's fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Calendar ID (internal)"
          },
          "uid": {
            "type": "string",
            "description": "Calendar UID (public)"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "color": {
            "type": "string",
            "description": "Assigned color"
          },
          "syncToken": {
            "type": "string",
            "description": "Current sync token"
          },
          "aclFol": {
            "type": "string",
            "description": "ACL info for folder itself"
          },
          "aclEle": {
            "type": "string",
            "description": "ACL info for folder elements"
          },
          "ownerUsername": {
            "type": "string",
            "description": "The owner profile's username"
          }
        },
        "required": [
          "id",
          "uid",
          "displayName",
          "syncToken",
          "aclFol",
          "aclEle",
          "ownerUsername"
        ]
      },
      "DavCalendarNew": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "color": {
            "type": "string",
            "description": "Color"
          }
        },
        "description": "Bean for carry addressbook's fields"
      },
      "DavCalObject": {
        "type": "object",
        "description": "Bean for carry calendar object's fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "CalObject ID (internal)"
          },
          "uid": {
            "type": "string",
            "description": "CalObject UID (public)"
          },
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "lastModified": {
            "type": "integer",
            "description": "Last modification time (unix timestamp)",
            "format": "int64"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          },
          "size": {
            "type": "integer",
            "description": "Size (in bytes) of calendar data"
          },
          "icalendar": {
            "type": "string",
            "description": "Calendar data (iCalendar format)"
          }
        }
      },
      "DavCalendarUpdate": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "New value for displayName"
          },
          "description": {
            "type": "string",
            "description": "New value for description"
          },
          "color": {
            "type": "string",
            "description": "New value for color"
          },
          "updatedFields": {
            "type": "array",
            "description": "Specifies which of above fields have been updated",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Bean for carry calendar's updateable fields"
      },
      "DavCalObjectNew": {
        "required": [
          "href"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "icalendar": {
            "type": "string",
            "description": "Calendar data (iCalendar format)"
          }
        },
        "description": "Bean for carry new calendar object's fields"
      },
      "DavCalObjectChanged": {
        "type": "object",
        "description": "Bean for carry fields of calendar object that have been changed",
        "properties": {
          "id": {
            "type": "string",
            "description": "CalObject ID (internal)"
          },
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          }
        },
        "required": [
          "id",
          "href",
          "etag"
        ]
      },
      "DavCalObjectsChanges": {
        "required": [
          "deleted",
          "inserted",
          "syncToken",
          "updated"
        ],
        "type": "object",
        "properties": {
          "syncToken": {
            "type": "string",
            "description": "Current sync token"
          },
          "inserted": {
            "type": "array",
            "description": "Items that have been inserted",
            "items": {
              "$ref": "#/components/schemas/DavCalObjectChanged"
            }
          },
          "updated": {
            "type": "array",
            "description": "Items that have been updated",
            "items": {
              "$ref": "#/components/schemas/DavCalObjectChanged"
            }
          },
          "deleted": {
            "type": "array",
            "description": "Items that have been deleted",
            "items": {
              "$ref": "#/components/schemas/DavCalObjectChanged"
            }
          }
        },
        "description": "Bean for carry calendar object collection changes"
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "auth-apikey-bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "API-Key Authentication: specifies the secret token (bearer and username are both required)"
      },
      "auth-apikey-username": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Username",
        "description": "API-Key Authentication: specifies target username (bearer and username are both required)"
      }
    },
...
  },
...
}